fixed constraints for db

+++ database/migrations/2024_09_26_053755_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->text('review_text');
-            $table->integer('rating');
+            $table->unsignedTinyInteger('rating');
             $table->foreignId('reviewer_id')->constrained('students')->onDelete('cascade');
             $table->foreignId('reviewee_id')->constrained('students')->onDelete('cascade');
             $table->foreignId('assessment_id')->constrained('assessments')->onDelete('cascade');
@@ -23,16 +23,17 @@ return new class extends Migration
         });
     }
 
+
     /**
      * Reverse the migrations.
      */
     public function down(): void
     {
-    
+
         DB::statement('PRAGMA foreign_keys = OFF;');
 
         Schema::dropIfExists('reviews');
-    
+
         // Re-enable foreign key checks
         DB::statement('PRAGMA foreign_keys = ON;');
     }
